requirements.txt
# vllm==0.6.0
pyyaml==6.0.2
pydantic==2.9.0
openai==1.44.0